Removed unnecessary cast
`ColumnName` has an implicit conversion to `int` so the equality operator does not need the cast.

+++ src/Engine/ColumnName.cs
@@ -166,7 +166,7 @@ namespace Dgt.Minesweeper.Engine
 
             return (hasColumnName, hasValue) switch
             {
-                (true, true) => (int)columnName! == value,
+                (true, true) => columnName! == value,
                 (true, false) => false,
                 (false, true) => false,
                 (false, false) => true
